+++ docs/CHANGELOG.md
@@ -14,12 +14,14 @@ Entries inside each section should be ordered by type:
 
 ## Catalog, Lambdas
 !-->
-# unreleased - YYYY-MM-DD
+# 6.0.0a4 - 2024-06-18
 ## Python API
 
+* [Added] New `quilt3.admin` API with more features (requires 1.53+ stack) ([#3990](https://github.com/quiltdata/quilt/pull/3990))
+* [Removed] `quilt3.admin` API ([#3990](https://github.com/quiltdata/quilt/pull/3990))
 * [Removed] Drop Python 3.8 support ([#3993](https://github.com/quiltdata/quilt/pull/3993))
-
-## CLI
+* [Fixed] If upload optimization during `push()` succeeds the checksum is calculated from local file instead of remote file ([#3968](https://github.com/quiltdata/quilt/pull/3968))
+* [Changed] Upload optimization check now tries to use S3 SHA-256 checksum and falls back to ETag ([#3968](https://github.com/quiltdata/quilt/pull/3968))
 
 ## Catalog, Lambdas
 * [Added] Add "Qurator" helper for summarizing files content using Bedrock API ([#3989](https://github.com/quiltdata/quilt/pull/3989))